Enhance SEO and site structure with new metadata and sitemap functionality
- Added site URL configuration to .env.example for better SEO management.
- Implemented dynamic metadata generation in various layouts to improve Open Graph and Twitter card support.
- Created sitemap and robots.txt generation for better search engine indexing.
- Refactored politician profile page to utilize new SEO utilities and improve data fetching.
- Introduced new API functions for fetching politician data and generating static parameters for pre-rendering.
