+++ frontend/.env.example
@@ -9,6 +9,9 @@ GOOGLE_CLIENT_SECRET=your_google_client_secret
 # Backend API URL
 NEXT_PUBLIC_API_URL=http://localhost:8000/api/v1
 
+# Canonical site URL (SEO: Open Graph, sitemap, canonical links). Defaults to NEXTAUTH_URL or localhost.
+NEXT_PUBLIC_SITE_URL=http://localhost:3000
+
 # Google Analytics 4 — leave empty to disable tracking
 NEXT_PUBLIC_GA_MEASUREMENT_ID=
 #Supabase
